refactor: ensure listing of clients and their respective assets
